# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Ask Arthur Threat Intelligence API
  version: "2.0.0"
  description: |
    Real-time Australian scam and fraud threat intelligence API.
    Powered by community-sourced scam reports analysed by AI.

    ## Authentication
    All endpoints require a Bearer token (API key) provided during onboarding.

    ## Rate Limits
    Three tiers are available. See [Pricing](https://askarthur.au/pricing) for details.

    | Tier | Daily Limit | RPM | Max Batch Size |
    |------|-------------|-----|----------------|
    | Free | 25 | 60 | 10 |
    | Pro | 100 | 60 | 100 |
    | Enterprise | 5,000 | 300 | 500 |

    ## Caching
    Most GET endpoints return `Cache-Control: public, s-maxage=300` headers.

    ## Response freshness
    All v1 success responses include a top-level `lastUpdated` field with an
    ISO-8601 timestamp indicating when the underlying data was queried. When
    a response is served from CDN cache, `lastUpdated` reflects the moment
    the cache entry was populated — use it to gauge how stale the data layer
    is, independent of the wall-clock time you receive the response. Error
    responses (4xx/5xx) do not include `lastUpdated`.

    ## Changelog
    - **v2.3.0**: All v1 success responses now include a `lastUpdated`
      ISO-8601 timestamp at the top level. Additive, backwards-compatible.
    - **v2.2.0**: `enrichmentData` now includes additional namespaces from external
...
      and `riskLevel`.
  contact:
    name: Ask Arthur
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key provided during onboarding. Keys are stored as SHA-256 hashes.
  schemas:
...
    StatsResponse:
      type: object
      required:
        - generated_at
        - verified_threats
        - top_scam_types_7d
      properties:
        generated_at:
          type: string
          format: date-time
        verified_threats:
          type: object
          properties:
            last_24h:
              type: integer
              example: 12
            last_7d:
              type: integer
              example: 87
            last_30d:
              type: integer
              example: 342
        top_scam_types_7d:
          type: array
          items:
            type: object
            properties:
              scam_type:
                type: string
              count:
                type: integer
...
    BatchLookupRequest:
      type: object
      required:
        - entities
      properties:
        entities:
          type: array
          maxItems: 100
          items:
            type: object
            required:
              - type
              - value
            properties:
              type:
                type: string
                enum:
                  [phone, email, url, domain, ip, crypto_wallet, bank_account]
              value:
                type: string
    BatchLookupResponse:
      type: object
      required:
        - total
        - found
        - results
      properties:
        total:
          type: integer
        found:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
              found:
                type: boolean
              entityId:
                type: integer
              reportCount:
                type: integer
              firstSeen:
                type: string
                format: date-time
              lastSeen:
                type: string
                format: date-time
              riskScore:
                type: integer
                minimum: 0
                maximum: 100
                description: "Composite risk score 0-100 (added in v2.1.0)"
              riskLevel:
                type: string
                enum: [LOW, MEDIUM, HIGH, CRITICAL, UNKNOWN]
                description: "Risk level derived from score (added in v2.1.0)"
    EntityDetailResponse:
      type: object
      required:
        - entity
        - summary
        - reports
      properties:
        entity:
          type: object
          properties:
            id:
              type: integer
            entityType:
              type: string
            normalizedValue:
              type: string
            reportCount:
              type: integer
            firstSeen:
              type: string
              format: date-time
            lastSeen:
              type: string
              format: date-time
            enrichmentStatus:
              type: string
              enum: [none, pending, in_progress, completed, failed]
              description: "Status of entity enrichment (added in v2.1.0)"
            enrichmentData:
              type: object
              description: "Enrichment intelligence data with namespaced results (added in v2.1.0)"
              additionalProperties: true
            riskScore:
              type: integer
              minimum: 0
              maximum: 100
              description: "Composite risk score 0-100 (added in v2.1.0)"
            riskLevel:
              type: string
              enum: [LOW, MEDIUM, HIGH, CRITICAL, UNKNOWN]
              description: "Risk level derived from score (added in v2.1.0)"
            riskFactors:
              type: object
              description: "Breakdown of risk scoring factors (added in v2.1.0)"
              additionalProperties: true
        summary:
          type: object
          properties:
            verdictDistribution:
              type: object
              additionalProperties:
                type: integer
            scamTypes:
              type: object
              additionalProperties:
                type: integer
            clusterIds:
              type: array
              items:
                type: integer
        reports:
          type: array
          items:
            $ref: "#/components/schemas/ReportSummary"
    ReportSummary:
      type: object
      properties:
        reportId:
          type: integer
        role:
          type: string
        extractionMethod:
          type: string
        verdict:
          type: string
          enum: [SAFE, SUSPICIOUS, HIGH_RISK]
        confidenceScore:
          type: number
        scamType:
          type: string
        channel:
          type: string
        impersonatedBrand:
          type: string
        region:
          type: string
        reportedAt:
          type: string
          format: date-time
    ClusterSummary:
      type: object
      properties:
        id:
          type: integer
        clusterType:
          type: string
          enum: [entity_overlap, text_similarity, brand_campaign, manual]
        primaryScamType:
          type: string
        primaryBrand:
          type: string
        memberCount:
          type: integer
        entityCount:
          type: integer
        status:
          type: string
          enum: [active, dormant, disrupted]
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
paths:
...
  /threats/stats:
    get:
      summary: Get aggregate threat statistics
      description: Returns high-level statistics about verified threats across time periods.
      operationId: getThreatStats
      tags: [Threats]
      responses:
        "200":
          description: Aggregate statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /threats/urls/lookup:
    get:
      summary: Look up a URL
      description: Check if a specific URL has been reported as a scam threat.
      operationId: lookupUrl
      tags: [Threats]
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: The URL to look up
      responses:
        "200":
          description: URL lookup result
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  normalizedUrl:
                    type: string
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /entities/batch:
    post:
      summary: Batch entity lookup
      description: |
        Check up to 100 entities in a single request. Returns match status and
        basic details for each entity.
      operationId: batchLookupEntities
      tags: [Entities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchLookupRequest"
      responses:
        "200":
          description: Batch lookup results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchLookupResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /entities/{id}:
    get:
      summary: Get entity details
      description: |
        Returns full details for an entity including linked reports,
        verdict distribution, scam type breakdown, and cluster membership.
      operationId: getEntityDetail
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The entity ID
      responses:
        "200":
          description: Entity details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityDetailResponse"
        "404":
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /clusters:
    get:
      summary: List scam clusters
      description: |
        Returns clusters of related scam reports grouped by shared entities.
        Clusters are automatically detected by the pipeline.
      operationId: listClusters
      tags: [Clusters]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: active
            enum: [active, dormant, disrupted]
          description: Filter by cluster status
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Maximum number of clusters to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
      responses:
        "200":
          description: List of clusters
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
                  clusters:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClusterSummary"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...